Update server.js
I added an extra argument called query. So, basically you pass the query you want into the function and it returns paginated results for that specific query. There might be a better way to do this. This would make the function more dynamic and powerful as it can now take in complex queries.